missions/mission2/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
paho-mqtt==1.6.1
